feat: Classroom Migration From SDK
