+++ apps/br_ipti_tag_app/lib/app/features/classroom/domain/usecases/create_classroom_usecase.dart
@@ -1,8 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:tag_sdk/src/core/defaults/usecase.dart';
-import 'package:tag_sdk/src/core/failures/failures.dart';
-import 'package:tag_sdk/src/features/classroom/domain/entities/classroom_create_entity.dart';
-import 'package:tag_sdk/src/features/classroom/domain/repositories/classroom_repository.dart';
+import 'package:tag_sdk/tag_sdk.dart';
 
 class CreateClassroomUsecase extends Usecase<bool, ClassroomCreateEntity> {
   CreateClassroomUsecase(this._repositoryClassroom);
