+++ apps/br_ipti_tag_app/lib/app/features/classroom/domain/usecases/delete_classroom_usecase.dart
@@ -1,7 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:tag_sdk/src/core/defaults/usecase.dart';
-import 'package:tag_sdk/src/core/failures/failures.dart';
-import 'package:tag_sdk/src/features/classroom/domain/repositories/classroom_repository.dart';
+import 'package:tag_sdk/tag_sdk.dart';
 
 class DeleteClassroomUsecase extends Usecase<bool, int> {
   DeleteClassroomUsecase(this._repositoryClassroom);
