+++ apps/br_ipti_tag_app/lib/app/features/classroom/domain/usecases/list_instructors_usecase.dart
@@ -1,9 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:tag_sdk/src/core/defaults/params.dart';
-import 'package:tag_sdk/src/core/defaults/usecase.dart';
-import 'package:tag_sdk/src/core/failures/failures.dart';
-import 'package:tag_sdk/src/features/teacher/domain/entities/instructor.dart';
-import 'package:tag_sdk/src/features/teacher/domain/repositories/instructor_repository.dart';
+import 'package:tag_sdk/tag_sdk.dart';
 
 class ListInstructorsUseCase implements Usecase<List<Instructor>, EmptyParams> {
   final InstructorRepository _repositoryInstructor;
