+++ apps/br_ipti_tag_app/lib/app/features/classroom/presentation/create/bloc/classroom_create_bloc.dart
@@ -2,6 +2,7 @@ import 'dart:developer';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tag_sdk/tag_sdk.dart';
+import '../../../domain/usecases/create_classroom_usecase.dart';
 import 'classroom_create_events.dart';
 import 'classroom_states.dart';
