+++ apps/br_ipti_tag_app/lib/app/features/classroom/presentation/list/bloc/classroom_list_bloc.dart
@@ -5,6 +5,7 @@ import 'package:br_ipti_tag_app/app/core/util/enums/status_fetch.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tag_sdk/tag_sdk.dart';
 
+import '../../../domain/usecases/list_classrooms_usecase.dart';
 import 'classroom_list_states.dart';
 
 class ClassroomListBloc extends Cubit<ClassroomListState> {
