+++ apps/br_ipti_tag_app/lib/app/features/classroom/presentation/update_delete/bloc/classroom_update_bloc.dart
@@ -1,6 +1,8 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tag_sdk/tag_sdk.dart';
 
+import '../../../domain/usecases/delete_classroom_usecase.dart';
+import '../../../domain/usecases/update_classroom_usecase.dart';
 import 'classroom_states.dart';
 import 'classroom_update_events.dart';
