+++ apps/br_ipti_tag_app/lib/app/features/classroom/presentation/update_delete/bloc/update_teacher/update_teacher_bloc.dart
@@ -2,6 +2,8 @@ import 'package:br_ipti_tag_app/app/features/classroom/presentation/update_delet
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tag_sdk/tag_sdk.dart';
 
+import '../../../../domain/usecases/list_instructors_teaching_data_usecase.dart';
+
 final initialState = UpdateTeacherStateLoading();
 
 class UpdateTeacherBloc extends Cubit<UpdateTeacherState> {
