+++ packages/tag_sdk/lib/src/features/auth/domain/repositories/auth_repository_impl.dart
@@ -2,8 +2,6 @@ import 'dart:developer';
 
 import 'package:dartz/dartz.dart';
 import 'package:tag_network/tag_network.dart';
-import 'package:tag_sdk/src/features/auth/data/datasources/local/auth_local_datasource.dart';
-import 'package:tag_sdk/src/features/auth/data/datasources/remote/auth_remote_datasource.dart';
 import 'package:tag_sdk/tag_sdk.dart';
 
 class AuthRepositoryImpl extends AuthRepository {
