+++ packages/tag_sdk/lib/src/features/classroom/classroom_module.dart
@@ -3,15 +3,6 @@ import 'package:tag_network/tag_network.dart';
 
 import 'data/datasource/classroom_datasource.dart';
 import 'data/repositories/classroom_repository_impl.dart';
-import 'domain/repositories/classroom_repository.dart';
-import 'domain/usecases/create_classroom_usecase.dart';
-import 'domain/usecases/create_instructor_teaching_data_usecase.dart';
-import 'domain/usecases/delete_classroom_usecase.dart';
-import 'domain/usecases/list_classrooms_usecase.dart';
-
-import 'domain/usecases/list_instructors_teaching_data_usecase.dart';
-import 'domain/usecases/update_classroom_usecase.dart';
-import 'domain/usecases/update_instructor_teaching_data_usecase.dart';
 
 class ClassroomSDKModule extends Module {
   @override
@@ -31,59 +22,6 @@ class ClassroomSDKModule extends Module {
       export: true,
     ),
 
-    // List
-    Bind.singleton(
-      (i) => ListClassroomsUsecase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
 
-    Bind.singleton(
-      (i) => ListInstructorsTeachingDataUseCase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
-
-    // Create
-    Bind.singleton(
-      (i) => CreateClassroomUsecase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
-    Bind.singleton(
-      (i) => CreateInstructorTeachingDataUseCase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
-
-    //UpdateDelete
-    Bind.singleton(
-      (i) => UpdateClassroomUsecase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
-    Bind.singleton(
-      (i) => DeleteClassroomUsecase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
-    Bind.singleton(
-      (i) => CreateInstructorTeachingDataUseCase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
-    Bind.singleton(
-      (i) => UpdateInstructorTeachingDataUseCase(
-        i.get<ClassroomRepository>(),
-      ),
-      export: true,
-    ),
   ];
 }
