+++ packages/tag_sdk/lib/src/features/classroom/data/model/list_instructors_teaching_data_params.dart
@@ -1,0 +1,16 @@
+class ListInstructorsTeachingDataParams {
+  final int classroomId;
+  final String perPage;
+  final String page;
+  const ListInstructorsTeachingDataParams({
+    required this.classroomId,
+    this.page = "1",
+    this.perPage = '100',
+  });
+
+  Map<String, dynamic> toJson() => <String, dynamic>{
+    "page": page,
+    "perPage": perPage,
+    "classroomId": classroomId,
+  };
+}
