+++ packages/tag_sdk/lib/src/features/classroom/domain/repositories/classroom_repository.dart
@@ -1,6 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:tag_sdk/src/core/api/classroom/get_classroom_endpoint.dart';
 import 'package:tag_sdk/src/core/failures/failures.dart';
+import 'package:tag_sdk/src/features/classroom/data/model/list_instructors_teaching_data_params.dart';
 
 import 'package:tag_sdk/src/features/classroom/domain/entities/classroom_create_entity.dart';
 import 'package:tag_sdk/src/features/classroom/domain/entities/classroom_entity.dart';
@@ -8,8 +9,6 @@ import 'package:tag_sdk/src/features/classroom/domain/entities/classroom_entity.
 import 'package:tag_sdk/src/features/classroom/domain/entities/instructor_teaching_data_create_entity.dart';
 import 'package:tag_sdk/src/features/classroom/domain/entities/instructors_teaching_data_entity.dart';
 import 'package:tag_sdk/src/features/classroom/domain/entities/update_instructor_teaching_data_entity.dart';
-
-import 'package:tag_sdk/src/features/classroom/domain/usecases/list_instructors_teaching_data_usecase.dart';
 
 abstract class ClassroomRepository {
   Future<Either<Failure, bool>> create(ClassroomCreateEntity classroom);
