+++ packages/tag_sdk/lib/src/features/features.dart
@@ -17,6 +17,7 @@ export '../core/failures/failures.dart';
 // export 'auth/domain/usecases/verify_auth_usecase.dart';
 export 'classroom/classroom_module.dart';
 export 'classroom/data/datasource/classroom_datasource.dart';
+export 'classroom/data/model/list_instructors_teaching_data_params.dart';
 export 'classroom/data/model/classroom_create_model.dart';
 export 'classroom/data/model/classroom_model.dart';
 export 'classroom/data/model/instructor_teaching_data_create_model.dart';
@@ -30,13 +31,13 @@ export 'classroom/domain/entities/instructor_teaching_data_create_entity.dart';
 export 'classroom/domain/entities/instructors_teaching_data_entity.dart';
 export 'classroom/domain/entities/update_instructor_teaching_data_entity.dart';
 export 'classroom/domain/repositories/classroom_repository.dart';
-export 'classroom/domain/usecases/create_classroom_usecase.dart';
-export 'classroom/domain/usecases/create_instructor_teaching_data_usecase.dart';
-export 'classroom/domain/usecases/delete_classroom_usecase.dart';
-export 'classroom/domain/usecases/list_classrooms_usecase.dart';
-export 'classroom/domain/usecases/list_instructors_teaching_data_usecase.dart';
-export 'classroom/domain/usecases/update_classroom_usecase.dart';
-export 'classroom/domain/usecases/update_instructor_teaching_data_usecase.dart';
+// export 'classroom/domain/usecases/create_classroom_usecase.dart';
+// export 'classroom/domain/usecases/create_instructor_teaching_data_usecase.dart';
+// export 'classroom/domain/usecases/delete_classroom_usecase.dart';
+// export 'classroom/domain/usecases/list_classrooms_usecase.dart';
+// export 'classroom/domain/usecases/list_instructors_teaching_data_params.dart';
+// export 'classroom/domain/usecases/update_classroom_usecase.dart';
+// export 'classroom/domain/usecases/update_instructor_teaching_data_usecase.dart';
 export 'edcenso_disciplines/data/datasources/edcenso_locations_local_datasource.dart';
 export 'edcenso_disciplines/data/datasources/edcenso_locations_remote_datasource.dart';
 export 'edcenso_disciplines/data/models/edcenso_discipline.dart';
