+++ packages/tag_sdk/test/coverage_helper_test.dart
@@ -7,14 +7,14 @@ import 'package:tag_sdk/src/features/classroom/domain/entities/instructors_teach
 import 'package:tag_sdk/src/features/classroom/domain/entities/classroom_create_entity.dart';
 import 'package:tag_sdk/src/features/classroom/domain/entities/instructor_teaching_data_create_entity.dart';
 import 'package:tag_sdk/src/features/classroom/domain/entities/update_instructor_teaching_data_entity.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/update_classroom_usecase.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/create_instructor_teaching_data_usecase.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/update_instructor_teaching_data_usecase.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/list_instructors_usecase.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/list_instructors_teaching_data_usecase.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/create_classroom_usecase.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/delete_classroom_usecase.dart';
-import 'package:tag_sdk/src/features/classroom/domain/usecases/list_classrooms_usecase.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/update_classroom_usecase.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/create_instructor_teaching_data_usecase.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/update_instructor_teaching_data_usecase.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/list_instructors_usecase.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/list_instructors_teaching_data_params.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/create_classroom_usecase.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/delete_classroom_usecase.dart';
+// import 'package:tag_sdk/src/features/classroom/domain/usecases/list_classrooms_usecase.dart';
 import 'package:tag_sdk/src/features/classroom/domain/repositories/classroom_repository.dart';
 import 'package:tag_sdk/src/features/classroom/classroom_module.dart';
 import 'package:tag_sdk/src/features/classroom/data/model/classroom_create_model.dart';
